requirements.txt
schedule==0.6.0
python-firebase == 1.2
beautifulsoup4==4.9.3
requests==2.24.0